Chore(EventBoardArea): 게시판 영역 모듈화
