+++ src/app/(main)/event/_components/board-area/index.ts
@@ -1,0 +1,1 @@
+export { BoardArea as EventBoardArea } from './BoardArea'
